requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
deprecation==2.1.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
ntgcalls==2.0.5
packaging==24.0
psutil==5.9.8
py-tgcalls==2.2.5
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.1
screeninfo==0.8.1
TgCrypto==1.2.5
yarl==1.9.4
